# removed unrolling in fit to check for performance