removed unrolling in fit to check for performance

```
BenchmarkLassoRegression-8         	       9	 122681315 ns/op	336586220 B/op	  100214 allocs/op
BenchmarkLassoAutoRegression-8     	       6	 192317451 ns/op	360238789 B/op	  119103 allocs/op
BenchmarkLassoSerialRegression-8   	       3	 373061806 ns/op	1155485824 B/op	  101260 allocs/op
BenchmarkOLSRegression-8           	      14	  80235524 ns/op	18036425 B/op	     140 allocs/op
```